created bcow game event
