+++ commands/tasks/unicom/dailyYYY.js
@@ -32,7 +32,7 @@ let charMap =  ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C",
 
 var secretkeyArray = function () {
   for (let e = [], t = charMap, i = 0; 5 > i; i++) {
-    for (let n = "", s = 0; 16 > s; s++) {
+    for (var n = "", s = 0; 16 > s; s++) {
       let a = Math.floor(62 * Math.random());
       n += t[a];
     }
@@ -43,7 +43,7 @@ var secretkeyArray = function () {
 
 var newjiamarr = () => {
   for (let e = [], k = "", t = charMap, i = 0x0; 0x4 > i; i++) {
-    for (let n = "", s = 0x0; 16 > s; s++) {
+    for (var n = "", s = 0x0; 16 > s; s++) {
       let a = Math.floor(0x3e * Math.random());
       n += t[a];
     }
