+++ commands/tasks/unicom/handlers/myPhone.js
@@ -1,0 +1,5 @@
+const useragent = (options) =>
+  `Mozilla/5.0 (Linux; Android 7.1.2; SM-G977N Build/LMY48Z; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/75.0.3770.143 Mobile Safari/537.36; unicom{version:android@8.0102,desmobile:${options.user}};devicetype{deviceBrand:samsung,deviceModel:SM-G977N};{yw_code:}`;
+module.exports = {
+  useragent,
+};
